style: createComment, createReview api 분리
