+++ src/api/comment/createComment.ts
@@ -1,6 +1,6 @@
 import axios from "@api/axiosInstance";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { ICreateComment, ICreateReview } from "@type/commentOrReview";
+import { ICreateComment } from "@type/commentOrReview";
 import { useRouter } from "next/router";
 
 // 자유 게시판 댓글
@@ -25,43 +25,7 @@ export const useCreateCommentMutation = () => {
       await queryClient.invalidateQueries({
         queryKey: ["postDetail", variables.postId],
       });
-
       router.push(`/posts/${variables.postId}`);
-    },
-    onError: (err) => {
-      console.error(err);
-    },
-    onSettled: () => {
-      console.log("완료");
-    },
-  });
-};
-
-// 교육 게시판 리뷰
-export const createReview = async (data: ICreateReview) => {
-  try {
-    await axios.post(`/review`, data);
-
-    return true;
-  } catch (err) {
-    return false;
-  }
-};
-
-// 교육 게시판 리뷰 Mutation
-export const useCreateReviewMutation = () => {
-  const router = useRouter();
-
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    mutationFn: createReview,
-    onSuccess: async (_, variables) => {
-      await queryClient.invalidateQueries({
-        queryKey: ["educationDetail", variables.educationId],
-      });
-
-      router.push(`/educations/${variables.educationId}`);
     },
     onError: (err) => {
       console.error(err);
